+++ CHANGELOG.md
@@ -10,7 +10,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 - optimized the cleanup service to be more error resistent
 - baseprice referenceAmount defaults to 1.0 instead of the item
+- use correct category position and image for translations
+- use all translations if no direct translation is found for categories
 - ignore case of object types when searching definitions
+- use voucher number as order item name
 
 ### Fixed
 - Fatal error when variation unit is not mapped